Account for bigger ten area size
The zone where the ten can be expanded significantly since 2D. I increased the dimensions of possibleTenSquare to accommodate that. Unfortunately, I couldn't find a 3D bot that will drop off samples, so I wasn't able to test all the use-cases. Also, in order to make these changes, I had to get rid of the old drill spot picking code, so the drill spot picking on this commit is awful.

In its current state, this code is not practical for submission (very low points at the moment), but it plays in the spirit of the game. If ZR ever decides to stop people from using stupid strategies, this is a good fallback. Also, there's lots of room for improvement here in terms of point scoring.